# openapi.yaml
openapi: 3.0.1
info:
  title: eCMR - API
  description: eCMR - API
  contact:
    name: |
      Open Logistics Foundation
    url: https://openlogisticsfoundation.org/
    email: info@openlogisticsfoundation.org
  version: v0.1
servers:
  - url: http://localhost:8081/api
    description: Generated server url
security:
  - OAUth2: []
paths:
...
  /external/ecmr/{ecmrId}/email:
    post:
      tags:
        - ECMR External
      summary: Send eCMR sharing token via email
      operationId: sendEmail
      parameters:
        - name: ecmrId
          in: path
          description: UUID of the eCMR
          required: true
          schema:
            type: string
            format: uuid
        - name: receiverEmail
          in: query
          description: Email address of the receiver
          required: true
          schema:
            type: string
            format: email
        - name: ecmrRole
          in: query
          description: Role of the user for the eCMR
          required: true
          schema:
            type: string
            enum:
              - Reader
              - Sender
              - Carrier
              - Consignee
      responses:
        "200":
          description: Email sent successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EcmrShareResponse"
        "404":
          description: eCMR not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EcmrShareResponse"
        "401":
          description: Unauthorized access
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EcmrShareResponse"
        "403":
          description: Forbidden access
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EcmrShareResponse"
        "400":
          description: Sealed eCMR not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EcmrShareResponse"
  /external/ecmr/import:
    post:
      tags:
        - ECMR External
      summary: "Import eCMR with ID, share token and url"
      operationId: importEcmrFromExternal
      parameters:
        - name: ecmrId
          in: query
          description: UUID of the eCMR
          required: true
          schema:
            type: string
            format: uuid
        - name: shareToken
          in: query
          description: Share token
          required: true
          schema:
            type: string
        - name: groupIds
          in: query
          description: Group IDs
          required: true
          schema:
            type: object
        - name: url
          in: query
          description: URL of the external instance
          required: true
          schema:
            type: string
        - name: groupId
          in: query
          required: true
          schema:
            type: array
            items:
              type: integer
              format: int64
      responses:
        "200":
          description: eCMR was imported successfully
        "401":
          description: Unauthorized access
        "403":
          description: Forbidden access
        "400":
          description: Share token is invalid
...
  /external/ecmr/{ecmrId}/export:
    get:
      tags:
        - ECMR External
      summary: Export eCMR as sealed document with ID and share token
      operationId: exportEcmrToExternal
      parameters:
        - name: ecmrId
          in: path
          description: UUID of the eCMR
          required: true
          schema:
            type: string
            format: uuid
        - name: shareToken
          in: query
          description: Share token
          required: true
          schema:
            type: string
      responses:
        default:
          description: The requested ecmr as sealed document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SealedDocument"
        "404":
          description: eCMR not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EcmrExportResult"
        "403":
          description: Forbidden access
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EcmrExportResult"
...
components:
  schemas:
    CarrierContactInformation:
      type: object
      properties:
        email:
          maxLength: 255
          minLength: 0
          type: string
        phone:
          maxLength: 32
          minLength: 0
          pattern: "^(\\+)?(?=.*\\d)[0-9\\s\\-().]+${1,32}"
          type: string
    CarrierCountryCode:
      type: object
      properties:
        region:
          maxLength: 60
          minLength: 2
          type: string
        value:
          maxLength: 2
          minLength: 2
          pattern: "^[A-Z]{2}$"
          type: string
    CarrierInformation:
      type: object
      properties:
        carrierNameCompany:
          type: string
        carrierNamePerson:
          maxLength: 60
          minLength: 2
          type: string
        carrierStreet:
          maxLength: 255
          minLength: 2
          type: string
        carrierPostcode:
          maxLength: 17
          minLength: 2
          type: string
        carrierCity:
          maxLength: 60
          minLength: 2
          type: string
        carrierCountryCode:
          $ref: "#/components/schemas/CarrierCountryCode"
        carrierLicensePlate:
          maxLength: 30
          minLength: 2
          type: string
        carrierContactInformation:
          $ref: "#/components/schemas/CarrierContactInformation"
    CarriersReservationsAndObservationsOnTakingOverTheGoods:
      type: object
      properties:
        carrierReservationsObservations:
          maxLength: 512
          minLength: 2
          type: string
        senderReservationsObservationsSignature:
          $ref: "#/components/schemas/Signature"
    CashOnDelivery:
      type: object
      properties:
        customCashOnDelivery:
          maximum: 999999
          minimum: 1
          type: integer
          format: int32
    ConsigneeContactInformation:
      type: object
      properties:
        email:
          maxLength: 255
          minLength: 0
          type: string
        phone:
          maxLength: 32
          minLength: 0
          pattern: "^(\\+)?(?=.*\\d)[0-9\\s\\-().]+${1,32}"
          type: string
    ConsigneeCountryCode:
      type: object
      properties:
        region:
          maxLength: 60
          minLength: 2
          type: string
        value:
          maxLength: 2
          minLength: 2
          pattern: "^[A-Z]{2}$"
          type: string
    ConsigneeInformation:
      type: object
      properties:
        consigneeNameCompany:
          type: string
        consigneeNamePerson:
          maxLength: 60
          minLength: 2
          type: string
        consigneePostcode:
          maxLength: 17
          minLength: 2
          type: string
        consigneeCity:
          maxLength: 60
          minLength: 2
          type: string
        consigneeCountryCode:
          $ref: "#/components/schemas/ConsigneeCountryCode"
        consigneeStreet:
          maxLength: 255
          minLength: 2
          type: string
        consigneeContactInformation:
          $ref: "#/components/schemas/ConsigneeContactInformation"
    CustomCharge:
      type: object
      properties:
        value:
          maximum: 99999
          minimum: 0
          type: number
          format: float
        currency:
          maxLength: 512
          minLength: 2
          type: string
        payer:
          type: string
          enum:
            - SENDER
            - CONSIGNEE
    DeliveryOfTheGoods:
      type: object
      properties:
        logisticsLocationCity:
          maxLength: 60
          minLength: 2
          type: string
        logisticsLocationOpeningHours:
          maxLength: 255
          minLength: 2
          type: string
    DocumentsHandedToCarrier:
      type: object
      properties:
        documentsRemarks:
          maxLength: 512
          minLength: 2
          type: string
    EcmrConsignment:
      type: object
      properties:
        senderInformation:
          $ref: "#/components/schemas/SenderInformation"
        multiConsigneeShipment:
          $ref: "#/components/schemas/MultiConsigneeShipment"
        consigneeInformation:
          $ref: "#/components/schemas/ConsigneeInformation"
        takingOverTheGoods:
          $ref: "#/components/schemas/TakingOverTheGoods"
        deliveryOfTheGoods:
          $ref: "#/components/schemas/DeliveryOfTheGoods"
        sendersInstructions:
          $ref: "#/components/schemas/SendersInstructions"
        carrierInformation:
          $ref: "#/components/schemas/CarrierInformation"
        successiveCarrierInformation:
          $ref: "#/components/schemas/SuccessiveCarrierInformation"
        carriersReservationsAndObservationsOnTakingOverTheGoods:
          $ref: "#/components/schemas/CarriersReservationsAndObservationsOnTakingOverTheGoods"
        documentsHandedToCarrier:
          $ref: "#/components/schemas/DocumentsHandedToCarrier"
        itemList:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        specialAgreementsSenderCarrier:
          $ref: "#/components/schemas/SpecialAgreementsSenderCarrier"
        toBePaidBy:
          $ref: "#/components/schemas/ToBePaidBy"
        otherUsefulParticulars:
          $ref: "#/components/schemas/OtherUsefulParticulars"
        cashOnDelivery:
          $ref: "#/components/schemas/CashOnDelivery"
        established:
          $ref: "#/components/schemas/Established"
        signatureOrStampOfTheSender:
          $ref: "#/components/schemas/SignatureOrStampOfTheSender"
        signatureOrStampOfTheCarrier:
          $ref: "#/components/schemas/SignatureOrStampOfTheCarrier"
        goodsReceived:
          $ref: "#/components/schemas/GoodsReceived"
        nonContractualPartReservedForTheCarrier:
          $ref: "#/components/schemas/NonContractualPartReservedForTheCarrier"
        referenceIdentificationNumber:
          $ref: "#/components/schemas/ReferenceIdentificationNumber"
    EcmrModel:
      required:
        - ecmrConsignment
      type: object
      properties:
        ecmrId:
          type: string
        ecmrConsignment:
          $ref: "#/components/schemas/EcmrConsignment"
        ecmrStatus:
          type: string
          enum:
            - NEW
            - LOADING
            - IN_TRANSPORT
            - ARRIVED_AT_DESTINATION
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        editedAt:
          type: string
          format: date-time
        editedBy:
          type: string
        originUrl:
          type: string
    Established:
      type: object
      properties:
        customEstablishedDate:
          type: string
          format: date-time
        customEstablishedIn:
          maxLength: 30
          minLength: 2
          type: string
    GoodsReceived:
      type: object
      properties:
        confirmedLogisticsLocationName:
          maxLength: 60
          minLength: 2
          type: string
        consigneeReservationsObservations:
          maxLength: 512
          minLength: 2
          type: string
        consigneeSignature:
          $ref: "#/components/schemas/Signature"
        consigneeSignatureDate:
          type: string
          format: date-time
        consigneeTimeOfArrival:
          type: string
          format: date-time
        consigneeTimeOfDeparture:
          type: string
          format: date-time
    GrossWeightInKg:
      type: object
      properties:
        supplyChainConsignmentItemGrossWeight:
          maximum: 99999
          minimum: 1
          type: integer
          format: int32
    Item:
      type: object
      properties:
        marksAndNos:
          $ref: "#/components/schemas/MarksAndNos"
        numberOfPackages:
          $ref: "#/components/schemas/NumberOfPackages"
        methodOfPacking:
          $ref: "#/components/schemas/MethodOfPacking"
        natureOfTheGoods:
          $ref: "#/components/schemas/NatureOfTheGoods"
        grossWeightInKg:
          $ref: "#/components/schemas/GrossWeightInKg"
        volumeInM3:
          $ref: "#/components/schemas/VolumeInM3"
    MarksAndNos:
      type: object
      properties:
        logisticsShippingMarksMarking:
          maxLength: 512
          minLength: 2
          type: string
        logisticsShippingMarksCustomBarcode:
          maxLength: 35
          minLength: 2
          type: string
    MethodOfPacking:
      type: object
      properties:
        logisticsPackageType:
          maxLength: 35
          minLength: 2
          type: string
    MultiConsigneeShipment:
      required:
        - isMultiConsigneeShipment
      type: object
      properties:
        isMultiConsigneeShipment:
          type: boolean
    NatureOfTheGoods:
      type: object
      properties:
        transportCargoIdentification:
          maxLength: 512
          minLength: 2
          type: string
    NonContractualPartReservedForTheCarrier:
      type: object
      properties:
        nonContractualCarrierRemarks:
          maxLength: 512
          minLength: 2
          type: string
    NumberOfPackages:
      type: object
      properties:
        logisticsPackageItemQuantity:
          maximum: 9999
          minimum: 0
          type: integer
          format: int32
    OtherUsefulParticulars:
      type: object
      properties:
        customParticulars:
          maxLength: 512
          minLength: 2
          type: string
    ReferenceIdentificationNumber:
      type: object
      properties:
        value:
          maxLength: 35
          minLength: 1
          type: string
    SenderContactInformation:
      type: object
      properties:
        email:
          maxLength: 255
          minLength: 0
          type: string
        phone:
          maxLength: 32
          minLength: 0
          pattern: "^(\\+)?(?=.*\\d)[0-9\\s\\-().]+${1,32}"
          type: string
    SenderCountryCode:
      type: object
      properties:
        region:
          maxLength: 60
          minLength: 2
          type: string
        value:
          maxLength: 2
          minLength: 2
          pattern: "^[A-Z]{2}$"
          type: string
    SenderInformation:
      type: object
      properties:
        senderNameCompany:
          type: string
        senderNamePerson:
          maxLength: 60
          minLength: 2
          type: string
        senderStreet:
          maxLength: 255
          minLength: 2
          type: string
        senderPostcode:
          maxLength: 17
          minLength: 2
          type: string
        senderCity:
          maxLength: 60
          minLength: 2
          type: string
        senderCountryCode:
          $ref: "#/components/schemas/SenderCountryCode"
        senderContactInformation:
          $ref: "#/components/schemas/SenderContactInformation"
    SendersInstructions:
      type: object
      properties:
        transportInstructionsDescription:
          maxLength: 512
          minLength: 2
          type: string
    Signature:
      type: object
      properties:
        type:
          type: string
        userName:
          type: string
        userCompany:
          type: string
        userStreet:
          type: string
        userPostCode:
          type: string
        userCity:
          type: string
        userCountry:
          type: string
        timestamp:
          type: string
          format: date-time
        data:
          type: string
    SignatureOrStampOfTheCarrier:
      type: object
      properties:
        carrierSignature:
          $ref: "#/components/schemas/Signature"
    SignatureOrStampOfTheSender:
      type: object
      properties:
        senderSignature:
          $ref: "#/components/schemas/Signature"
    SpecialAgreementsSenderCarrier:
      type: object
      properties:
        customSpecialAgreement:
          maxLength: 255
          minLength: 2
          type: string
    SuccessiveCarrierContactInformation:
      type: object
      properties:
        email:
          maxLength: 255
          minLength: 0
          type: string
        phone:
          maxLength: 32
          minLength: 0
          pattern: "^(\\+)?(?=.*\\d)[0-9\\s\\-().]+${1,32}"
          type: string
    SuccessiveCarrierCountryCode:
      type: object
      properties:
        region:
          maxLength: 60
          minLength: 2
          type: string
        value:
          maxLength: 2
          minLength: 2
          pattern: "^[A-Z]{2}$"
          type: string
    SuccessiveCarrierInformation:
      type: object
      properties:
        successiveCarrierCity:
          maxLength: 60
          minLength: 2
          type: string
        successiveCarrierCountryCode:
          $ref: "#/components/schemas/SuccessiveCarrierCountryCode"
        successiveCarrierNameCompany:
          type: string
        successiveCarrierNamePerson:
          maxLength: 60
          minLength: 2
          type: string
        successiveCarrierPostcode:
          maxLength: 17
          minLength: 2
          type: string
        successiveCarrierSignature:
          $ref: "#/components/schemas/Signature"
        successiveCarrierSignatureDate:
          type: string
          format: date-time
        successiveCarrierStreet:
          maxLength: 255
          minLength: 2
          type: string
        successiveCarrierContactInformation:
          $ref: "#/components/schemas/SuccessiveCarrierContactInformation"
    TakingOverTheGoods:
      type: object
      properties:
        takingOverTheGoodsPlace:
          maxLength: 60
          minLength: 2
          type: string
        logisticsTimeOfArrivalDateTime:
          type: string
          format: date-time
        logisticsTimeOfDepartureDateTime:
          type: string
          format: date-time
    ToBePaidBy:
      type: object
      properties:
        customChargeCarriage:
          $ref: "#/components/schemas/CustomCharge"
        customChargeSupplementary:
          $ref: "#/components/schemas/CustomCharge"
        customChargeCustomsDuties:
          $ref: "#/components/schemas/CustomCharge"
        customChargeOther:
          $ref: "#/components/schemas/CustomCharge"
    VolumeInM3:
      type: object
      properties:
        supplyChainConsignmentItemGrossVolume:
          maximum: 9999
          minimum: 1
          type: integer
          format: int32
...
    Group:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        children:
          type: array
          items:
            $ref: "#/components/schemas/Group"
...
    EcmrShareResponse:
      type: object
      properties:
        result:
          type: string
          enum:
            - SharedInternal
            - SharedExternal
            - ErrorInternalUserHasNoGroup
        group:
          $ref: "#/components/schemas/Group"
...
    Metadata:
      type: object
      properties:
        sealer:
          type: string
        timestamp:
          type: string
          format: date-time
    SealedDocument:
      required:
        - seal
        - sealedEcmr
      type: object
      properties:
        sealedEcmr:
          $ref: "#/components/schemas/SealedEcmr"
        seal:
          type: string
        precedingSeal:
          type: string
    SealedEcmr:
      required:
        - ecmr
        - metadata
      type: object
      properties:
        metadata:
          $ref: "#/components/schemas/Metadata"
        ecmr:
          $ref: "#/components/schemas/EcmrModel"
    EcmrExportResult:
      type: object
      properties:
        sealedDocument:
          $ref: "#/components/schemas/SealedDocument"
        ecmrRole:
          type: string
          enum:
            - Reader
            - Sender
            - Carrier
            - Consignee
...
  securitySchemes:
    OAUth2:
      type: oauth2
      flows:
        authorizationCode: {}